+++ elements/pl_matrix_input/pl_matrix_input.py
@@ -2,6 +2,7 @@ import prairielearn as pl
 import lxml.html
 from html import escape
 import numpy as np
+import random
 import math
 import chevron
 
@@ -18,6 +19,11 @@ def render(element_html, element_index, data):
     element = lxml.html.fragment_fromstring(element_html)
     name = pl.get_string_attrib(element, 'answers_name')
     label = pl.get_string_attrib(element, 'label', None)
+
+    if '_pl_matrix_input_format' in data['submitted_answers']:
+        format_type = data['submitted_answers']['_pl_matrix_input_format'].get(name, 'matlab')
+    else:
+        format_type = 'matlab'
 
     if data['panel'] == 'question':
         editable = data['editable']
@@ -70,7 +76,10 @@ def render(element_html, element_index, data):
         html_params = {'submission': True, 'label': label, 'parse_error': parse_error}
         if parse_error is None:
             a_sub = np.array(data['submitted_answers'][name])
-            html_params['a_sub'] = pl.numpy_to_matlab(a_sub, ndigits=12, wtype='g')
+            if format_type == 'matlab':
+                html_params['a_sub'] = pl.numpy_to_matlab(a_sub, ndigits=12, wtype='g')
+            else:
+                html_params['a_sub'] = str(np.array(a_sub).tolist())
         else:
             raw_submitted_answer = data['raw_submitted_answers'].get(name, None)
             if raw_submitted_answer is not None:
@@ -105,18 +114,24 @@ def render(element_html, element_index, data):
                 rtol = pl.get_float_attrib(element, 'rtol', 1e-5)
                 atol = pl.get_float_attrib(element, 'atol', 1e-8)
                 # FIXME: render correctly with respect to rtol and atol
-                a_tru = pl.numpy_to_matlab(a_tru, ndigits=12, wtype='g')
+                matlab_data = pl.numpy_to_matlab(a_tru, ndigits=12, wtype='g')
+                python_data = str(np.array(a_tru).tolist())
             elif comparison == 'sigfig':
                 digits = pl.get_integer_attrib(element, 'digits', 2)
-                a_tru = pl.numpy_to_matlab_sf(a_tru, ndigits=digits)
+                matlab_data = pl.numpy_to_matlab_sf(a_tru, ndigits=digits)
+                python_data = pl.string_from_2darray_sf(a_tru, ndigits=digits)
             elif comparison == 'decdig':
                 digits = pl.get_integer_attrib(element, 'digits', 2)
-                a_tru = pl.numpy_to_matlab(a_tru, ndigits=digits, wtype='f')
+                matlab_data = pl.numpy_to_matlab(a_tru, ndigits=digits, wtype='f')
+                python_data = str(np.array(a_tru).round(digits).tolist())
             else:
                 raise ValueError('method of comparison "%s" is not valid (must be "relabs", "sigfig", or "decdig")' % comparison)
 
-            # FIXME: render correctly with respect to method of comparison
-            html_params = {'answer': True, 'label': label, 'a_tru': a_tru}
+            html_params = {'answer': True, 'label': label, 'matlab_data': matlab_data, 'python_data': python_data, 'element_index': element_index}
+            if format_type == 'matlab':
+                html_params['default_is_matlab'] = True
+            else:
+                html_params['default_is_python'] = True
             with open('pl_matrix_input.mustache', 'r') as f:
                 html = chevron.render(f, html_params).strip()
         else:
@@ -145,14 +160,19 @@ def parse(element_html, element_index, data):
     a_sub = a_sub.replace(u'\u2212', '-')
 
     # Convert submitted answer to numpy array (return parse_error on failure)
-    (a_sub_parsed, parse_error) = pl.matlab_to_numpy(a_sub)
+    (a_sub_parsed, info) = pl.string_to_2darray(a_sub)
     if a_sub_parsed is None:
-        data['format_errors'][name] = parse_error
+        data['format_errors'][name] = info['format_error']
         data['submitted_answers'][name] = None
         return data
 
     # Replace submitted answer with numpy array
     data['submitted_answers'][name] = a_sub_parsed.tolist()
+
+    # Store format type
+    if '_pl_matrix_input_format' not in data['submitted_answers']:
+        data['submitted_answers']['_pl_matrix_input_format'] = {}
+    data['submitted_answers']['_pl_matrix_input_format'][name] = info['format_type']
 
     return data
 
@@ -211,5 +231,47 @@ def grade(element_html, element_index, data):
         data['partial_scores'][name] = {'score': 1, 'weight': weight}
     else:
         data['partial_scores'][name] = {'score': 0, 'weight': weight}
+
+    return data
+
+
+def test(element_html, element_index, data):
+    element = lxml.html.fragment_fromstring(element_html)
+    name = pl.get_string_attrib(element, 'answers_name')
+    weight = pl.get_integer_attrib(element, 'weight', 1)
+
+    a_tru = np.array(data['correct_answers'][name])
+
+    result = random.choices(['correct', 'incorrect', 'invalid'], [5, 5, 1])[0]
+    if random.choice([True, False]):
+        # matlab
+        if result == 'correct':
+            data['raw_submitted_answers'][name] = pl.numpy_to_matlab(a_tru, ndigits=12, wtype='g')
+            data['partial_scores'][name] = {'score': 1, 'weight': weight}
+        elif result == 'incorrect':
+            data['raw_submitted_answers'][name] = pl.numpy_to_matlab(a_tru + random.random(), ndigits=12, wtype='g')
+            data['partial_scores'][name] = {'score': 0, 'weight': weight}
+        elif result == 'invalid':
+            # FIXME: add more invalid expressions, make text of format_errors
+            # correct, and randomize
+            data['raw_submitted_answers'][name] = '[1, 2, 3]'
+            data['format_errors'][name] = 'invalid'
+        else:
+            raise Exception('invalid result: %s' % result)
+    else:
+        # python
+        if result == 'correct':
+            data['raw_submitted_answers'][name] = str(np.array(a_tru).tolist())
+            data['partial_scores'][name] = {'score': 1, 'weight': weight}
+        elif result == 'incorrect':
+            data['raw_submitted_answers'][name] = str(np.array(a_tru + random.random()).tolist())
+            data['partial_scores'][name] = {'score': 0, 'weight': weight}
+        elif result == 'invalid':
+            # FIXME: add more invalid expressions, make text of format_errors
+            # correct, and randomize
+            data['raw_submitted_answers'][name] = '[[1, 2, 3], [4, 5]]'
+            data['format_errors'][name] = 'invalid'
+        else:
+            raise Exception('invalid result: %s' % result)
 
     return data